Retracing claim about setting deliverymode on individual message not working..
I used the jmsMsg.setJMSDeliveryMode(deliveryMode), which is for broker
code, instead of producer.send(jmsMsg, deliveryMode, pri, ttl). When
using the latter, it works just the same as setting it on the producer.

The main point still stands, though: [Transactional, non-Persistent]
with JournalDiskSyncStragey.ALWAYS is just as slow [Transactional,
Persistent] and [non-Transactional, Persistent].